Broke up microstate test into seperate test files
